requirements.txt
absl-py==1.3.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
astroid==2.12.12
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
beartype==0.10.4
beautifulsoup4==4.11.1
bleach==5.0.1
blis==0.7.9
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
confection==0.0.3
contourpy==1.0.5
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
Django==4.0.5
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.10.0
fastjsonschema==2.16.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
flatbuffers==22.11.23
flet==0.1.37
fonttools==4.37.4
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
greenlet==1.1.3.post0
grpcio==1.50.0
h5py==3.7.0
idna==3.3
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.10
jsonschema==4.14.0
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.5
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
keras==2.11.0
kiwisolver==1.4.4
langcodes==3.3.0
lazy-object-proxy==1.8.0
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==0.8.4
murmurhash==1.0.9
nbclassic==0.4.3
nbclient==0.6.6
nbconvert==6.5.3
nbformat==5.4.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.2
oauthlib==3.2.1
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.2
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
preshed==3.0.8
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.6
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pylint==2.15.5
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python3-linkedin==1.0.2
pytz==2022.2.1
pywin32==304
pywinpty==2.0.7
pyzmq==23.2.1
regex==2022.7.9
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smart-open==5.2.1
sniffio==1.2.0
soupsieve==2.3.2.post1
spacy==3.4.2
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==1.4.41
sqlparse==0.4.2
srsly==2.4.5
stack-data==0.4.0
termcolor==2.1.1
terminado==0.15.0
thinc==8.1.5
threadpoolctl==3.1.0
tinycss2==1.1.1
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typer==0.4.2
typing_extensions==4.4.0
tzdata==2022.1
urllib3==1.26.11
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1